Cardano.Analysis.API.Ground

Documentation

newtype Hash

Constructors

Hash

Fields

unHash :: ShortText

newtype Host

Constructors

Host

Fields

unHost :: ShortText

data LogObjectSource

Constructors

LogObjectSourceJSON JsonLogfile

LogObjectSourceSQLite FilePath

LogObjectSourceOther FilePath

newtype Count (a :: k)

Constructors

Count

Fields

unCount :: Int

newtype FieldName

Constructors

FieldName

Fields

unFieldName :: Text

newtype TId

Constructors

TId

Fields

unTId :: ShortText

newtype EpochSlot

Constructors

EpochSlot

Fields

unEpochSlot :: Word64

newtype EpochSafeInt

Constructors

EpochSafeInt

Fields

unEpochSafeInt :: Word64

data HostDeduction

Constructors

HostFromLogfilename

newtype InputDir

Constructors

InputDir

Fields

unInputDir :: FilePath

newtype JsonLogfile

Constructors

JsonLogfile

Fields

unJsonLogfile :: FilePath

newtype JsonInputFile (a :: Type)

Constructors

JsonInputFile

Fields

unJsonInputFile :: FilePath

newtype JsonOutputFile (a :: Type)

Constructors

JsonOutputFile

Fields

unJsonOutputFile :: FilePath

newtype GnuplotOutputFile

Constructors

GnuplotOutputFile

Fields

unGnuplotOutputFile :: FilePath

newtype OrgOutputFile

Constructors

OrgOutputFile

Fields

unOrgOutputFile :: FilePath

newtype TextInputFile

Constructors

TextInputFile

Fields

unTextInputFile :: FilePath

newtype TextOutputFile

Constructors

TextOutputFile

Fields

unTextOutputFile :: FilePath

newtype CsvOutputFile

Constructors

CsvOutputFile

Fields

unCsvOutputFile :: FilePath

newtype SqliteOutputFile

Constructors

SqliteOutputFile

Fields

unSqliteOutputFile :: FilePath

newtype OutputFile

Constructors

OutputFile

Fields

unOutputFile :: FilePath

shortHash :: Hash -> Text

countMap :: Map a b -> Count a

countList :: (a -> Bool) -> [a] -> Count a

countLists :: (a -> Bool) -> [[a]] -> Count a

countListAll :: [a] -> Count a

countListsAll :: [[a]] -> Count a

unsafeCoerceCount :: Count a -> Count b

logObjectSourceFile :: LogObjectSource -> FilePath

toLogObjectSource :: FilePath -> LogObjectSource

readJsonData :: FromJSON a => JsonInputFile a -> (Text -> b) -> ExceptT b IO a

readJsonDataIO :: FromJSON a => JsonInputFile a -> IO (Either String a)

optInputDir :: String -> String -> Parser InputDir

optJsonLogfile :: String -> String -> Parser JsonLogfile

optLogObjectSource :: String -> String -> Parser LogObjectSource

argJsonLogfile :: Parser JsonLogfile

optJsonInputFile :: String -> String -> Parser (JsonInputFile a)

optJsonOutputFile :: String -> String -> Parser (JsonOutputFile a)

optGnuplotOutputFile :: String -> String -> Parser GnuplotOutputFile

optTextInputFile :: String -> String -> Parser TextInputFile

optTextOutputFile :: String -> String -> Parser TextOutputFile

optCsvOutputFile :: String -> String -> Parser CsvOutputFile

optSqliteOutputFile :: String -> String -> Parser SqliteOutputFile

optOutputFile :: String -> String -> Parser OutputFile

pSlotNo :: String -> String -> Parser SlotNo

optWord :: String -> String -> Word64 -> Parser Word64

optString :: String -> String -> Parser String

hostFromLogfilename :: JsonLogfile -> Host

hostDeduction :: HostDeduction -> JsonLogfile -> Host

dumpObject :: ToJSON a => String -> a -> JsonOutputFile a -> ExceptT Text IO ()

dumpObjects :: ToJSON a => String -> [a] -> JsonOutputFile [a] -> ExceptT Text IO ()

dumpAssociatedObjects :: ToJSON a => String -> [(LogObjectSource, a)] -> ExceptT Text IO ()

dumpAssociatedObjectsWith :: (a -> ByteString) -> String -> [(LogObjectSource, a)] -> ExceptT Text IO ()

dumpAssociatedObjectsPretty :: ToJSON a => String -> [(LogObjectSource, a)] -> ExceptT Text IO ()

readAssociatedObjects :: forall a. FromJSON a => String -> [JsonLogfile] -> ExceptT Text IO [(LogObjectSource, a)]

dumpAssociatedObjectStreams :: ToJSON a => String -> [(LogObjectSource, [a])] -> ExceptT Text IO ()

dumpText :: String -> [Text] -> TextOutputFile -> ExceptT Text IO ()

dumpAssociatedTextStreams :: String -> [(LogObjectSource, [Text])] -> ExceptT Text IO ()

module Data.DataDomain

data UTCTime

This is the simplest representation of UTC. It consists of the day number, and a time offset from midnight. Note that if a day has a leap second added to it, it will have 86401 seconds.

data NominalDiffTime

This is a length of time, as measured by UTC. It has a precision of 10^-12 s.

Conversion functions such as fromInteger and realToFrac will treat it as seconds. For example, (0.010 :: NominalDiffTime) corresponds to 10 milliseconds.

It has a precision of one picosecond (= 10^-12 s). Enumeration functions will treat it as picoseconds.

It ignores leap-seconds, so it's not necessarily a fixed amount of clock time. For instance, 23:00 UTC + 2 hours of NominalDiffTime = 01:00 UTC (+ 1 day), regardless of whether a leap-second intervened.

newtype BlockNo

The 0-based index of the block in the blockchain. BlockNo is <= SlotNo and is only equal at slot N if there is a block for every slot where N <= SlotNo.

Constructors

BlockNo

Fields

unBlockNo :: Word64

newtype EpochNo

An epoch, i.e. the number of the epoch.

Constructors

EpochNo

Fields

unEpochNo :: Word64

newtype SlotNo

The 0-based index for the Ourboros time slot.

Constructors

SlotNo

Fields

unSlotNo :: Word64

Orphan instances

Data BlockNo

Instance details

Methods

gfoldl :: (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> BlockNo -> c BlockNo

gunfold :: (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> Constr -> c BlockNo

toConstr :: BlockNo -> Constr

dataTypeOf :: BlockNo -> DataType

dataCast1 :: Typeable t => (forall d. Data d => c (t d)) -> Maybe (c BlockNo)

dataCast2 :: Typeable t => (forall d e. (Data d, Data e) => c (t d e)) -> Maybe (c BlockNo)

gmapT :: (forall b. Data b => b -> b) -> BlockNo -> BlockNo

gmapQl :: (r -> r' -> r) -> r -> (forall d. Data d => d -> r') -> BlockNo -> r

gmapQr :: forall r r'. (r' -> r -> r) -> r -> (forall d. Data d => d -> r') -> BlockNo -> r

gmapQ :: (forall d. Data d => d -> u) -> BlockNo -> [u]

gmapQi :: Int -> (forall d. Data d => d -> u) -> BlockNo -> u

gmapM :: Monad m => (forall d. Data d => d -> m d) -> BlockNo -> m BlockNo

gmapMp :: MonadPlus m => (forall d. Data d => d -> m d) -> BlockNo -> m BlockNo

gmapMo :: MonadPlus m => (forall d. Data d => d -> m d) -> BlockNo -> m BlockNo

Data SlotNo

Instance details

Methods

gfoldl :: (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> SlotNo -> c SlotNo

gunfold :: (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> Constr -> c SlotNo

toConstr :: SlotNo -> Constr

dataTypeOf :: SlotNo -> DataType

dataCast1 :: Typeable t => (forall d. Data d => c (t d)) -> Maybe (c SlotNo)

dataCast2 :: Typeable t => (forall d e. (Data d, Data e) => c (t d e)) -> Maybe (c SlotNo)

gmapT :: (forall b. Data b => b -> b) -> SlotNo -> SlotNo

gmapQl :: (r -> r' -> r) -> r -> (forall d. Data d => d -> r') -> SlotNo -> r

gmapQr :: forall r r'. (r' -> r -> r) -> r -> (forall d. Data d => d -> r') -> SlotNo -> r

gmapQ :: (forall d. Data d => d -> u) -> SlotNo -> [u]

gmapQi :: Int -> (forall d. Data d => d -> u) -> SlotNo -> u

gmapM :: Monad m => (forall d. Data d => d -> m d) -> SlotNo -> m SlotNo

gmapMp :: MonadPlus m => (forall d. Data d => d -> m d) -> SlotNo -> m SlotNo

gmapMo :: MonadPlus m => (forall d. Data d => d -> m d) -> SlotNo -> m SlotNo

Real BlockNo

Instance details

Methods

toRational :: BlockNo -> Rational

Real SlotNo

Instance details

Methods

toRational :: SlotNo -> Rational

Divisible BlockNo

Instance details

Methods

divide :: BlockNo -> Double -> BlockNo

fromDouble :: Double -> BlockNo

Divisible SlotNo

Instance details

Methods

divide :: SlotNo -> Double -> SlotNo

fromDouble :: Double -> SlotNo