Cardano.TopHandler

Documentation

toplevelExceptionHandler :: IO a -> IO a

An exception handler to use for a program top level, as an alternative to the default top level handler provided by GHC.

Use like:

main :: IO ()
main = toplevelExceptionHandler $ do
  ...