Cardano.Unlog.LogObjectDB

Description

This module contains the schema in which LogObjectss are stored to and retrieved from an SQLite database.

logObjectToSql provides the mapping for LOBody with its constructors of varying arity to different DB tables and columns. toLOBodyConverters implements the reverse, mapping the columns of a result row to a LOBody value.

The AsSqlData type class provides lower level data marshalling.

Documentation

class AsSQLData x where

Minimal complete definition

toSqlData, fromSqlData

Methods

toSqlData :: x -> SQLData

fromSqlData :: SQLData -> x

fromSqlDataWithArgs :: [SQLData] -> x

data SummaryDB

Constructors

SummaryDB

Fields

sdbName :: Host

sdbLines :: Int

sdbFirstAt :: UTCTime

sdbLastAt :: UTCTime

sdbCreated :: UTCTime

type SQLRunnable = (SQL, [SQLData])

data SQLSelect6Cols

A select statement to retrieve log objects from the DB where the result set has exactly 6 columns, with all 4 argument columns being nullable: at (timestamp) | LogObject constructor (text) | int arg 1 | int arg 2 | float arg | text or blob arg

mkSQLSelectFrom :: Text -> Maybe Text -> Maybe Text -> Maybe Text -> Maybe Text -> Maybe Text -> SQLSelect6Cols

Smart constructor to ensure 6 columns in the result set. The at column is assumed to exist in table.

sqlAppend :: SQLSelect6Cols -> Text -> SQLSelect6Cols

Append (possibly WHERE clause filters) to an existing select statement

sqlOrdered :: [SQLSelect6Cols] -> SQL

Union of SELECTs, with the result rows ordered by timestamp

type TraceFreqs = Map Text Int

sqlToLogObject :: SummaryDB -> [SQLData] -> LogObject

logObjectToSql :: LogObject -> Maybe SQLRunnable

errorToSql :: String -> String -> SQLRunnable

summaryToSql :: SummaryDB -> SQLRunnable

traceFreqsToSql :: TraceFreqs -> [SQLRunnable]

createSchema :: [SQL]

createLMetricsIfNotExists :: SQL

runSqlRunnable :: SQLRunnable -> SQLite [[SQLData]]

fromSqlDataPair :: (AsSQLData a, AsSQLData b) => [SQLData] -> (a, b)

toSqlDataPair :: (AsSQLData a, AsSQLData b) => (a, b) -> [SQLData]

allLOBodyConstructors :: Set Text

knownLOBodyConstructors :: Set Text