Cardano.Util

Documentation

error :: HasCallStack => [Char] -> a

error stops execution and displays an error message.

head :: HasCallStack => [a] -> a

Warning: This is a partial function, it throws an error on empty lists. Use pattern matching, uncons or listToMaybe instead. Consider refactoring to use Data.List.NonEmpty.

\(\mathcal{O}(1)\). Extract the first element of a list, which must be non-empty.

To disable the warning about partiality put {-# OPTIONS_GHC -Wno-x-partial -Wno-unrecognised-warning-flags #-} at the top of the file. To disable it throughout a package put the same options into ghc-options section of Cabal file. To disable it in GHCi put :set -Wno-x-partial -Wno-unrecognised-warning-flags into ~/.ghci config file. See also the migration guide.

Examples

Expand

>>> head [1, 2, 3]
1

>>> head [1..]
1

>>> head []
*** Exception: Prelude.head: empty list

last :: HasCallStack => [a] -> a

\(\mathcal{O}(n)\). Extract the last element of a list, which must be finite and non-empty.

WARNING: This function is partial. Consider using unsnoc instead.

Examples

Expand

>>> last [1, 2, 3]
3

>>> last [1..]
* Hangs forever *

>>> last []
*** Exception: Prelude.last: empty list

type String = [Char]

String is an alias for a list of characters.

String constants in Haskell are values of type String. That means if you write a string literal like "hello world", it will have the type [Char], which is the same as String.

Note: You can ask the compiler to automatically infer different types with the -XOverloadedStrings language extension, for example "hello world" :: Text. See IsString for more information.

Because String is just a list of characters, you can use normal list functions to do basic string manipulation. See Data.List for operations on lists.

Performance considerations

Expand

[Char] is a relatively memory-inefficient type. It is a linked list of boxed word-size characters, internally it looks something like:

╭─────┬───┬──╮  ╭─────┬───┬──╮  ╭─────┬───┬──╮  ╭────╮
│ (:) │   │ ─┼─>│ (:) │   │ ─┼─>│ (:) │   │ ─┼─>│ [] │
╰─────┴─┼─┴──╯  ╰─────┴─┼─┴──╯  ╰─────┴─┼─┴──╯  ╰────╯
        v               v               v
       'a'             'b'             'c'

The String "abc" will use 5*3+1 = 16 (in general 5n+1) words of space in memory.

Furthermore, operations like (++) (string concatenation) are O(n) (in the left argument).

For historical reasons, the base library uses String in a lot of places for the conceptual simplicity, but library code dealing with user-data should use the text package for Unicode text, or the the bytestring package for binary data.

(.!=) :: Parser (Maybe a) -> a -> Parser a

(.:) :: FromJSON a => Object -> Key -> Parser a

(.:?) :: FromJSON a => Object -> Key -> Parser (Maybe a)

withObject :: String -> (Object -> Parser a) -> Value -> Parser a

object :: [Pair] -> Value

class FromJSON a where

Minimal complete definition

Nothing

Methods

parseJSON :: Value -> Parser a

parseJSONList :: Value -> Parser [a]

omittedField :: Maybe a

type Object = KeyMap Value

data Value

Constructors

Object !Object

Array !Array

String !Text

Number !Scientific

Bool !Bool

Null

class ToJSON a where

Minimal complete definition

Nothing

Methods

toJSON :: a -> Value

toEncoding :: a -> Encoding

toJSONList :: [a] -> Value

toEncodingList :: [a] -> Encoding

omitField :: a -> Bool

high :: Interval v -> v

low :: Interval v -> v

point :: v -> Interval v

data Interval v

Constructors

Interval v v

unI :: I a -> a

newtype I a

Constructors

I a

fn :: forall {k} f (a :: k) f'. (f a -> f' a) -> (f -.-> f') a

fn_2 :: forall {k} f (a :: k) f' f''. (f a -> f' a -> f'' a) -> (f -.-> (f' -.-> f'')) a

fn_3 :: forall {k} f (a :: k) f' f'' f'''. (f a -> f' a -> f'' a -> f''' a) -> (f -.-> (f' -.-> (f'' -.-> f'''))) a

fn_4 :: forall {k} f (a :: k) f' f'' f''' f''''. (f a -> f' a -> f'' a -> f''' a -> f'''' a) -> (f -.-> (f' -.-> (f'' -.-> (f''' -.-> f'''')))) a

hcfoldMap :: forall {k} {l} h c (xs :: l) m proxy f. (HTraverse_ h, AllN h c xs, Monoid m) => proxy c -> (forall (a :: k). c a => f a -> m) -> h f xs -> m

hcfor :: forall {l} h c (xs :: l) g proxy f. (HSequence h, AllN h c xs, Applicative g) => proxy c -> h f xs -> (forall a. c a => f a -> g a) -> g (h I xs)

hcfor_ :: forall {k} {l} h c (xs :: l) g proxy f. (HTraverse_ h, AllN h c xs, Applicative g) => proxy c -> h f xs -> (forall (a :: k). c a => f a -> g ()) -> g ()

hcliftA :: forall {k} {l} h c (xs :: l) proxy f f'. (AllN (Prod h) c xs, HAp h) => proxy c -> (forall (a :: k). c a => f a -> f' a) -> h f xs -> h f' xs

hcliftA2 :: forall {k} {l} h c (xs :: l) proxy f f' f''. (AllN (Prod h) c xs, HAp h, HAp (Prod h)) => proxy c -> (forall (a :: k). c a => f a -> f' a -> f'' a) -> Prod h f xs -> h f' xs -> h f'' xs

hcliftA3 :: forall {k} {l} h c (xs :: l) proxy f f' f'' f'''. (AllN (Prod h) c xs, HAp h, HAp (Prod h)) => proxy c -> (forall (a :: k). c a => f a -> f' a -> f'' a -> f''' a) -> Prod h f xs -> Prod h f' xs -> h f'' xs -> h f''' xs

hcmap :: forall {k} {l} h c (xs :: l) proxy f f'. (AllN (Prod h) c xs, HAp h) => proxy c -> (forall (a :: k). c a => f a -> f' a) -> h f xs -> h f' xs

hctraverse :: forall {l} h c (xs :: l) g proxy f. (HSequence h, AllN h c xs, Applicative g) => proxy c -> (forall a. c a => f a -> g a) -> h f xs -> g (h I xs)

hczipWith :: forall {k} {l} h c (xs :: l) proxy f f' f''. (AllN (Prod h) c xs, HAp h, HAp (Prod h)) => proxy c -> (forall (a :: k). c a => f a -> f' a -> f'' a) -> Prod h f xs -> h f' xs -> h f'' xs

hczipWith3 :: forall {k} {l} h c (xs :: l) proxy f f' f'' f'''. (AllN (Prod h) c xs, HAp h, HAp (Prod h)) => proxy c -> (forall (a :: k). c a => f a -> f' a -> f'' a -> f''' a) -> Prod h f xs -> Prod h f' xs -> h f'' xs -> h f''' xs

hfromI :: forall {l1} {k2} {l2} h1 (f :: k2 -> Type) (xs :: l1) (ys :: l2) h2. (AllZipN (Prod h1) (LiftedCoercible I f) xs ys, HTrans h1 h2) => h1 I xs -> h2 f ys

hliftA :: forall {k} {l} h (xs :: l) f f'. (SListIN (Prod h) xs, HAp h) => (forall (a :: k). f a -> f' a) -> h f xs -> h f' xs

hliftA2 :: forall {k} {l} h (xs :: l) f f' f''. (SListIN (Prod h) xs, HAp h, HAp (Prod h)) => (forall (a :: k). f a -> f' a -> f'' a) -> Prod h f xs -> h f' xs -> h f'' xs

hliftA3 :: forall {k} {l} h (xs :: l) f f' f'' f'''. (SListIN (Prod h) xs, HAp h, HAp (Prod h)) => (forall (a :: k). f a -> f' a -> f'' a -> f''' a) -> Prod h f xs -> Prod h f' xs -> h f'' xs -> h f''' xs

hmap :: forall {k} {l} h (xs :: l) f f'. (SListIN (Prod h) xs, HAp h) => (forall (a :: k). f a -> f' a) -> h f xs -> h f' xs

hsequence :: forall {l} h (xs :: l) f. (SListIN h xs, SListIN (Prod h) xs, HSequence h, Applicative f) => h f xs -> f (h I xs)

hsequenceK :: forall {k} {l} h (xs :: l) f a. (SListIN h xs, SListIN (Prod h) xs, Applicative f, HSequence h) => h (K (f a) :: k -> Type) xs -> f (h (K a :: k -> Type) xs)

htoI :: forall {k1} {l1} {l2} h1 (f :: k1 -> Type) (xs :: l1) (ys :: l2) h2. (AllZipN (Prod h1) (LiftedCoercible f I) xs ys, HTrans h1 h2) => h1 f xs -> h2 I ys

hzipWith :: forall {k} {l} h (xs :: l) f f' f''. (SListIN (Prod h) xs, HAp h, HAp (Prod h)) => (forall (a :: k). f a -> f' a -> f'' a) -> Prod h f xs -> h f' xs -> h f'' xs

hzipWith3 :: forall {k} {l} h (xs :: l) f f' f'' f'''. (SListIN (Prod h) xs, HAp h, HAp (Prod h)) => (forall (a :: k). f a -> f' a -> f'' a -> f''' a) -> Prod h f xs -> Prod h f' xs -> h f'' xs -> h f''' xs

fn_5 :: forall {k} f0 (a :: k) f1 f2 f3 f4 f5. (f0 a -> f1 a -> f2 a -> f3 a -> f4 a -> f5 a) -> (f0 -.-> (f1 -.-> (f2 -.-> (f3 -.-> (f4 -.-> f5))))) a

ap_NP :: forall {k} (f :: k -> Type) (g :: k -> Type) (xs :: [k]). NP (f -.-> g) xs -> NP f xs -> NP g xs

coerce_NP :: forall {k1} {k2} (f :: k1 -> Type) (g :: k2 -> Type) (xs :: [k1]) (ys :: [k2]). AllZip (LiftedCoercible f g) xs ys => NP f xs -> NP g ys

collapse_NP :: forall {k} a (xs :: [k]). NP (K a :: k -> Type) xs -> [a]

cpure_NP :: forall {k} c (xs :: [k]) proxy f. All c xs => proxy c -> (forall (a :: k). c a => f a) -> NP f xs

ctraverse'_NP :: forall {k} c proxy (xs :: [k]) f f' g. (All c xs, Applicative g) => proxy c -> (forall (a :: k). c a => f a -> g (f' a)) -> NP f xs -> g (NP f' xs)

ctraverse__NP :: forall {k} c proxy (xs :: [k]) f g. (All c xs, Applicative g) => proxy c -> (forall (a :: k). c a => f a -> g ()) -> NP f xs -> g ()

hd :: forall {k} f (x :: k) (xs :: [k]). NP f (x ': xs) -> f x

map_NP' :: forall {k} f g (xs :: [k]). (forall (a :: k). f a -> g a) -> NP f xs -> NP g xs

npToSListI :: forall {k} (a :: k -> Type) (xs :: [k]) r. NP a xs -> (SListI xs => r) -> r

pure_NP :: forall {k} (xs :: [k]) f. SListI xs => (forall (a :: k). f a) -> NP f xs

singletonNP :: forall {k} f (x :: k). f x -> NP f '[x]

tl :: forall {k} (f :: k -> Type) (x :: k) (xs :: [k]). NP f (x ': xs) -> NP f xs

trans_NP :: forall {k1} {k2} c (xs :: [k1]) (ys :: [k2]) proxy f g. AllZip c xs ys => proxy c -> (forall (x :: k1) (y :: k2). c x y => f x -> g y) -> NP f xs -> NP g ys

traverse__NP :: forall {k} (xs :: [k]) f g. (SListI xs, Applicative g) => (forall (a :: k). f a -> g ()) -> NP f xs -> g ()

ap_NS :: forall {k} (f :: k -> Type) (g :: k -> Type) (xs :: [k]). NP (f -.-> g) xs -> NS f xs -> NS g xs

cexpand_NS :: forall {k} c proxy f (xs :: [k]). All c xs => proxy c -> (forall (x :: k). c x => f x) -> NS f xs -> NP f xs

coerce_NS :: forall {k1} {k2} (f :: k1 -> Type) (g :: k2 -> Type) (xs :: [k1]) (ys :: [k2]). AllZip (LiftedCoercible f g) xs ys => NS f xs -> NS g ys

collapse_NS :: forall {k} a (xs :: [k]). NS (K a :: k -> Type) xs -> a

ctraverse'_NS :: forall {k} c proxy (xs :: [k]) f f' g. (All c xs, Functor g) => proxy c -> (forall (a :: k). c a => f a -> g (f' a)) -> NS f xs -> g (NS f' xs)

expand_NS :: forall {k} (xs :: [k]) f. SListI xs => (forall (x :: k). f x) -> NS f xs -> NP f xs

index_NS :: forall {k} (f :: k -> Type) (xs :: [k]). NS f xs -> Int

injections :: forall {k} (xs :: [k]) (f :: k -> Type). SListI xs => NP (Injection f xs) xs

partition_NS :: forall {k} (xs :: [k]) (f :: k -> Type). SListI xs => [NS f xs] -> NP ([] :.: f) xs

sequence_NS' :: forall {k} (xs :: [k]) f (g :: k -> Type). Functor f => NS (f :.: g) xs -> f (NS g xs)

trans_NS :: forall {k1} {k2} proxy c f g (xs :: [k1]) (ys :: [k2]). AllZip c xs ys => proxy c -> (forall (x :: k1) (y :: k2). c x y => f x -> g y) -> NS f xs -> NS g ys

unZ :: forall {k} f (x :: k). NS f '[x] -> f x

newtype ((f :: k -> Type) -.-> (g :: k -> Type)) (a :: k)

Constructors

Fn

Fields

apFn :: f a -> g a

type family CollapseTo (h :: (k -> Type) -> l -> Type) x

class (Prod (Prod h) ~ Prod h, HPure (Prod h)) => HAp (h :: (k -> Type) -> l -> Type) where

Methods

hap :: forall (f :: k -> Type) (g :: k -> Type) (xs :: l). Prod h (f -.-> g) xs -> h f xs -> h g xs

class UnProd (Prod h) ~ h => HApInjs (h :: (k -> Type) -> l -> Type) where

Methods

hapInjs :: forall (xs :: l) (f :: k -> Type). SListIN h xs => Prod h f xs -> [h f xs]

class HCollapse (h :: (k -> Type) -> l -> Type) where

Methods

hcollapse :: forall (xs :: l) a. SListIN h xs => h (K a :: k -> Type) xs -> CollapseTo h a

class HExpand (h :: (k -> Type) -> l -> Type) where

Methods

hexpand :: forall (xs :: l) f. SListIN (Prod h) xs => (forall (x :: k). f x) -> h f xs -> Prod h f xs

hcexpand :: forall c (xs :: l) proxy f. AllN (Prod h) c xs => proxy c -> (forall (x :: k). c x => f x) -> h f xs -> Prod h f xs

class HIndex (h :: (k -> Type) -> l -> Type) where

Methods

hindex :: forall (f :: k -> Type) (xs :: l). h f xs -> Int

class HPure (h :: (k -> Type) -> l -> Type) where

Methods

hpure :: forall (xs :: l) f. SListIN h xs => (forall (a :: k). f a) -> h f xs

hcpure :: forall c (xs :: l) proxy f. AllN h c xs => proxy c -> (forall (a :: k). c a => f a) -> h f xs

class HAp h => HSequence (h :: (k -> Type) -> l -> Type) where

Methods

hsequence' :: forall (xs :: l) f (g :: k -> Type). (SListIN h xs, Applicative f) => h (f :.: g) xs -> f (h g xs)

hctraverse' :: forall c (xs :: l) g proxy f f'. (AllN h c xs, Applicative g) => proxy c -> (forall (a :: k). c a => f a -> g (f' a)) -> h f xs -> g (h f' xs)

htraverse' :: forall (xs :: l) g f f'. (SListIN h xs, Applicative g) => (forall (a :: k). f a -> g (f' a)) -> h f xs -> g (h f' xs)

class ((Same h1 :: (k2 -> Type) -> l2 -> Type) ~ h2, (Same h2 :: (k1 -> Type) -> l1 -> Type) ~ h1) => HTrans (h1 :: (k1 -> Type) -> l1 -> Type) (h2 :: (k2 -> Type) -> l2 -> Type) where

Methods

htrans :: forall c (xs :: l1) (ys :: l2) proxy f g. AllZipN (Prod h1) c xs ys => proxy c -> (forall (x :: k1) (y :: k2). c x y => f x -> g y) -> h1 f xs -> h2 g ys

hcoerce :: forall (f :: k1 -> Type) (g :: k2 -> Type) (xs :: l1) (ys :: l2). AllZipN (Prod h1) (LiftedCoercible f g) xs ys => h1 f xs -> h2 g ys

class HTraverse_ (h :: (k -> Type) -> l -> Type) where

Methods

hctraverse_ :: forall c (xs :: l) g proxy f. (AllN h c xs, Applicative g) => proxy c -> (forall (a :: k). c a => f a -> g ()) -> h f xs -> g ()

htraverse_ :: forall (xs :: l) g f. (SListIN h xs, Applicative g) => (forall (a :: k). f a -> g ()) -> h f xs -> g ()

type family Prod (h :: (k -> Type) -> l -> Type) :: (k -> Type) -> l -> Type

type family Same (h :: (k1 -> Type) -> l1 -> Type) :: (k2 -> Type) -> l2 -> Type

type family UnProd (h :: (k -> Type) -> l -> Type) :: (k -> Type) -> l -> Type

data NP (f :: k -> Type) (xs :: [k]) where

Constructors

Nil :: forall {k} (f :: k -> Type). NP f ('[] :: [k])

(:*) :: forall {k} (f :: k -> Type) (x :: k) (xs1 :: [k]). !(f x) -> !(NP f xs1) -> NP f (x ': xs1)

type Injection (f :: k -> Type) (xs :: [k]) = f -.-> (K (NS f xs) :: k -> Type)

data NS (f :: k -> Type) (xs :: [k]) where

Constructors

Z :: forall {k} (f :: k -> Type) (x :: k) (xs1 :: [k]). !(f x) -> NS f (x ': xs1)

S :: forall {k} (f :: k -> Type) (xs1 :: [k]) (x :: k). !(NS f xs1) -> NS f (x ': xs1)

chunksOf :: Int -> [e] -> [[e]]

addUTCTime :: NominalDiffTime -> UTCTime -> UTCTime

addUTCTime a b = a + b

diffUTCTime :: UTCTime -> UTCTime -> NominalDiffTime

diffUTCTime a b = a - b

data NominalDiffTime

This is a length of time, as measured by UTC. It has a precision of 10^-12 s.

Conversion functions such as fromInteger and realToFrac will treat it as seconds. For example, (0.010 :: NominalDiffTime) corresponds to 10 milliseconds.

It has a precision of one picosecond (= 10^-12 s). Enumeration functions will treat it as picoseconds.

It ignores leap-seconds, so it's not necessarily a fixed amount of clock time. For instance, 23:00 UTC + 2 hours of NominalDiffTime = 01:00 UTC (+ 1 day), regardless of whether a leap-second intervened.

data UTCTime

This is the simplest representation of UTC. It consists of the day number, and a time offset from midnight. Note that if a day has a leap second added to it, it will have 86401 seconds.

Constructors

UTCTime

Fields

utctDay :: Day

the day

utctDayTime :: DiffTime

the time from midnight, 0 <= t < 86401s (because of leap-seconds)

module Data.Time.Clock.POSIX

fromSMaybe :: a -> StrictMaybe a -> a

Same as fromMaybe

data StrictMaybe a

Constructors

SNothing

SJust !a

(&&&) :: Arrow a => a b c -> a b c' -> a b (c, c') infixr 3

Fanout: send the input to both argument arrows and combine their output.

The default definition may be overridden with a more efficient version if desired.

    ╭───────╮ c
  b │ ┌─ f ─┼───>
>───┼─┤     │
    │ └─ g ─┼───>
    ╰───────╯ c'

(***) :: Arrow a => a b c -> a b' c' -> a (b, b') (c, c') infixr 3

Split the input between the two argument arrows and combine their output. Note that this is in general not a functor.

The default definition may be overridden with a more efficient version if desired.

  b ╭─────╮ b'
>───┼─ f ─┼───>
>───┼─ g ─┼───>
  c ╰─────╯ c'

(<|>) :: Alternative f => f a -> f a -> f a infixl 3

An associative binary operation

forConcurrently :: Traversable t => t a -> (a -> IO b) -> IO (t b)

forConcurrently_ :: Foldable f => f a -> (a -> IO b) -> IO ()

mapConcurrently :: Traversable t => (a -> IO b) -> t a -> IO (t b)

mapConcurrently_ :: Foldable f => (a -> IO b) -> f a -> IO ()

firstExceptT :: forall (m :: Type -> Type) x y a. Functor m => (x -> y) -> ExceptT x m a -> ExceptT y m a

newExceptT :: m (Either x a) -> ExceptT x m a

module Ouroboros.Consensus.Util.Time

printf :: PrintfType r => String -> r

Format a variable number of arguments with the C-style formatting string.

>>> printf "%s, %d, %.4f" "hello" 123 pi
hello, 123, 3.1416

The return value is either String or (IO a) (which should be (IO ()), but Haskell's type system makes this hard).

The format string consists of ordinary characters and conversion specifications, which specify how to format one of the arguments to printf in the output string. A format specification is introduced by the % character; this character can be self-escaped into the format string using %%. A format specification ends with a format character that provides the primary information about how to format the value. The rest of the conversion specification is optional. In order, one may have flag characters, a width specifier, a precision specifier, and type-specific modifier characters.

Unlike C printf(3), the formatting of this printf is driven by the argument type; formatting is type specific. The types formatted by printf "out of the box" are:

Integral types, including Char

String

RealFloat types

printf is also extensible to support other types: see below.

A conversion specification begins with the character %, followed by zero or more of the following flags:

-      left adjust (default is right adjust)
+      always use a sign (+ or -) for signed conversions
space  leading space for positive numbers in signed conversions
0      pad with zeros rather than spaces
#      use an \"alternate form\": see below

When both flags are given, - overrides 0 and + overrides space. A negative width specifier in a * conversion is treated as positive but implies the left adjust flag.

The "alternate form" for unsigned radix conversions is as in C printf(3):

%o           prefix with a leading 0 if needed
%x           prefix with a leading 0x if nonzero
%X           prefix with a leading 0X if nonzero
%b           prefix with a leading 0b if nonzero
%[eEfFgG]    ensure that the number contains a decimal point

Any flags are followed optionally by a field width:

num    field width
*      as num, but taken from argument list

The field width is a minimum, not a maximum: it will be expanded as needed to avoid mutilating a value.

Any field width is followed optionally by a precision:

.num   precision
.      same as .0
.*     as num, but taken from argument list

Negative precision is taken as 0. The meaning of the precision depends on the conversion type.

Integral    minimum number of digits to show
RealFloat   number of digits after the decimal point
String      maximum number of characters

The precision for Integral types is accomplished by zero-padding. If both precision and zero-pad are given for an Integral field, the zero-pad is ignored.

Any precision is followed optionally for Integral types by a width modifier; the only use of this modifier being to set the implicit size of the operand for conversion of a negative operand to unsigned:

hh     Int8
h      Int16
l      Int32
ll     Int64
L      Int64

The specification ends with a format character:

c      character               Integral
d      decimal                 Integral
o      octal                   Integral
x      hexadecimal             Integral
X      hexadecimal             Integral
b      binary                  Integral
u      unsigned decimal        Integral
f      floating point          RealFloat
F      floating point          RealFloat
g      general format float    RealFloat
G      general format float    RealFloat
e      exponent format float   RealFloat
E      exponent format float   RealFloat
s      string                  String
v      default format          any type

The "%v" specifier is provided for all built-in types, and should be provided for user-defined type formatters as well. It picks a "best" representation for the given type. For the built-in types the "%v" specifier is converted as follows:

c      Char
u      other unsigned Integral
d      other signed Integral
g      RealFloat
s      String

Mismatch between the argument types and the format string, as well as any other syntactic or semantic errors in the format string, will cause an exception to be thrown at runtime.

Note that the formatting for RealFloat types is currently a bit different from that of C printf(3), conforming instead to showEFloat, showFFloat and showGFloat (and their alternate versions showFFloatAlt and showGFloatAlt). This is hard to fix: the fixed versions would format in a backward-incompatible way. In any case the Haskell behavior is generally more sensible than the C behavior. A brief summary of some key differences:

Haskell printf never uses the default "6-digit" precision used by C printf.

Haskell printf treats the "precision" specifier as indicating the number of digits after the decimal point.

Haskell printf prints the exponent of e-format numbers without a gratuitous plus sign, and with the minimum possible number of digits.

Haskell printf will place a zero after a decimal point when possible.

catSMaybes :: [SMaybe a] -> [a]

count :: (a -> Bool) -> [a] -> Int

foldEmpty :: r -> ([a] -> r) -> [a] -> r

fromRUTCTime :: RUTCTime -> UTCTime

intersectIntv :: Ord a => [Interval a] -> Interval a

intvDurationSec :: Interval UTCTime -> NominalDiffTime

isSJust :: SMaybe a -> Bool

isSNothing :: SMaybe a -> Bool

lazySMaybe :: SMaybe a -> Maybe a

mapAndUnzip :: (a -> (b, c)) -> [a] -> ([b], [c])

mapConcurrentlyPure :: NFData b => (a -> b) -> [a] -> IO [b]

mapHead :: (a -> a) -> [a] -> [a]

mapLast :: (a -> a) -> [a] -> [a]

mapSMaybe :: (a -> StrictMaybe b) -> [a] -> [b]

mapSMaybeFB :: (b -> r -> r) -> (a -> StrictMaybe b) -> a -> r -> r

nChunksEachOf :: Int -> Int -> Text -> [Text]

norm2Tuple :: ((a, b), c) -> (a, (b, c))

progress :: MonadIO m => String -> F -> m ()

progressLock :: Lock

redistribute :: (a, (b, c)) -> ((a, b), (a, c))

renderIntv :: (a -> Text) -> Interval a -> Text

replaceExtension :: FilePath -> String -> FilePath

roundUTCTimeDay :: UTCTime -> UTCTime

roundUTCTimeSec :: UTCTime -> UTCTime

sequenceConcurrentlyChunksOf :: Int -> [IO a] -> IO [a]

showText :: Show a => a -> Text

smaybe :: b -> (a -> b) -> StrictMaybe a -> b

spans :: (a -> Bool) -> [a] -> [Vector a]

strictMaybe :: Maybe a -> SMaybe a

toDouble :: Real a => a -> Double

toRUTCTime :: UTCTime -> RUTCTime

unionIntv :: Ord a => [Interval a] -> Interval a

unsafeNominalToUTC :: NominalDiffTime -> UTCTime

unsafeUTCToNominal :: UTCTime -> NominalDiffTime

utcTimeDeltaSec :: UTCTime -> UTCTime -> Int

withTimingInfo :: MonadIO m => String -> m a -> m a

zeroUTCTime :: UTCTime

A tweaked version of UTCTime that is able to have more instances. Structurally equivalent to difftime from zeroUTCTime

data F

Constructors

R String

RNoCR String

Q String

L [String]

ToJSON a => J a

newtype RUTCTime

Constructors

RUTCTime

Fields

unRUTCTime :: NominalDiffTime

type SMaybe a = StrictMaybe a

Orphan instances

Functor Interval

Instance details

Methods

fmap :: (a -> b) -> Interval a -> Interval b

(<$) :: a -> Interval b -> Interval a

FromJSON a => FromJSON (Interval a)

Instance details

Methods

parseJSON :: Value -> Parser (Interval a)

parseJSONList :: Value -> Parser [Interval a]

omittedField :: Maybe (Interval a)

FromJSON a => FromJSON (I a)

Instance details

Methods

parseJSON :: Value -> Parser (I a)

parseJSONList :: Value -> Parser [I a]

omittedField :: Maybe (I a)

ToJSON a => ToJSON (Interval a)

Instance details

Methods

toJSON :: Interval a -> Value

toEncoding :: Interval a -> Encoding

toJSONList :: [Interval a] -> Value

toEncodingList :: [Interval a] -> Encoding

omitField :: Interval a -> Bool

ToJSON a => ToJSON (I a)

Instance details

Methods

toJSON :: I a -> Value

toEncoding :: I a -> Encoding

toJSONList :: [I a] -> Value

toEncodingList :: [I a] -> Encoding

omitField :: I a -> Bool

Data a => Data (SMaybe a)

Instance details

Methods

gfoldl :: (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> SMaybe a -> c (SMaybe a)

gunfold :: (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> Constr -> c (SMaybe a)

toConstr :: SMaybe a -> Constr

dataTypeOf :: SMaybe a -> DataType

dataCast1 :: Typeable t => (forall d. Data d => c (t d)) -> Maybe (c (SMaybe a))

dataCast2 :: Typeable t => (forall d e. (Data d, Data e) => c (t d e)) -> Maybe (c (SMaybe a))

gmapT :: (forall b. Data b => b -> b) -> SMaybe a -> SMaybe a

gmapQl :: (r -> r' -> r) -> r -> (forall d. Data d => d -> r') -> SMaybe a -> r

gmapQr :: forall r r'. (r' -> r -> r) -> r -> (forall d. Data d => d -> r') -> SMaybe a -> r

gmapQ :: (forall d. Data d => d -> u) -> SMaybe a -> [u]

gmapQi :: Int -> (forall d. Data d => d -> u) -> SMaybe a -> u

gmapM :: Monad m => (forall d. Data d => d -> m d) -> SMaybe a -> m (SMaybe a)

gmapMp :: MonadPlus m => (forall d. Data d => d -> m d) -> SMaybe a -> m (SMaybe a)

gmapMo :: MonadPlus m => (forall d. Data d => d -> m d) -> SMaybe a -> m (SMaybe a)