Data.Accum

Documentation

data Accum a b

Constructors

Accum

Fields

aUpdate :: !(NominalDiffTime -> a -> a -> b)

aPrevStamp :: !UTCTime

aPrevValue :: !a

aCurrent :: !b

mkAccum :: a -> b -> (NominalDiffTime -> a -> a -> b) -> Accum a b

divAccum :: Accum a Word64 -> Word64 -> Accum a Word64

Given an Accum, produce one that returns results downscaled by N.

mulAccum :: Accum a Word64 -> Word64 -> Accum a Word64

Given an Accum, produce one that returns results upscaled by N.

updateAccum :: UTCTime -> a -> Accum a b -> Accum a b

mkAccumNew :: Accum Word64 Word64

Just store the latest value.

mkAccumDelta :: Accum Word64 Word64

Simply compute the increase.

mkAccumTicksShare :: Accum Word64 Word64

Interpret values as centiseconds (ticks, in Linux), and compute ratio of elapsed time, in percents.