Cardano.Util

Documentation

type String = [Char]

String is an alias for a list of characters.

String constants in Haskell are values of type String. That means if you write a string literal like "hello world", it will have the type [Char], which is the same as String.

Note: You can ask the compiler to automatically infer different types with the -XOverloadedStrings language extension, for example "hello world" :: Text. See IsString for more information.

Because String is just a list of characters, you can use normal list functions to do basic string manipulation. See Data.List for operations on lists.

Performance considerations

Expand

[Char] is a relatively memory-inefficient type. It is a linked list of boxed word-size characters, internally it looks something like:

╭─────┬───┬──╮  ╭─────┬───┬──╮  ╭─────┬───┬──╮  ╭────╮
│ (:) │   │ ─┼─>│ (:) │   │ ─┼─>│ (:) │   │ ─┼─>│ [] │
╰─────┴─┼─┴──╯  ╰─────┴─┼─┴──╯  ╰─────┴─┼─┴──╯  ╰────╯
        v               v               v
       'a'             'b'             'c'

The String "abc" will use 5*3+1 = 16 (in general 5n+1) words of space in memory.

Furthermore, operations like (++) (string concatenation) are O(n) (in the left argument).

For historical reasons, the base library uses String in a lot of places for the conceptual simplicity, but library code dealing with user-data should use the text package for Unicode text, or the the bytestring package for binary data.

error :: forall (r :: RuntimeRep) (a :: TYPE r). HasCallStack => [Char] -> a

error stops execution and displays an error message.

head :: HasCallStack => [a] -> a

\(\mathcal{O}(1)\). Extract the first element of a list, which must be non-empty.

>>> head [1, 2, 3]
1
>>> head [1..]
1
>>> head []
*** Exception: Prelude.head: empty list

WARNING: This function is partial. You can use case-matching, uncons or listToMaybe instead.

last :: HasCallStack => [a] -> a

\(\mathcal{O}(n)\). Extract the last element of a list, which must be finite and non-empty.

>>> last [1, 2, 3]
3
>>> last [1..]
* Hangs forever *
>>> last []
*** Exception: Prelude.last: empty list

WARNING: This function is partial. You can use reverse with case-matching, uncons or listToMaybe instead.

data Value

Constructors

Object !Object

Array !Array

String !Text

Number !Scientific

Bool !Bool

Null

type Object = KeyMap Value

class FromJSON a where

Minimal complete definition

Nothing

Methods

parseJSON :: Value -> Parser a

parseJSONList :: Value -> Parser [a]

omittedField :: Maybe a

class ToJSON a where

Minimal complete definition

Nothing

Methods

toJSON :: a -> Value

toEncoding :: a -> Encoding

toJSONList :: [a] -> Value

toEncodingList :: [a] -> Encoding

omitField :: a -> Bool

object :: [Pair] -> Value

(.:) :: FromJSON a => Object -> Key -> Parser a

(.!=) :: Parser (Maybe a) -> a -> Parser a

(.:?) :: FromJSON a => Object -> Key -> Parser (Maybe a)

withObject :: String -> (Object -> Parser a) -> Value -> Parser a

data Interval v

A closed interval. The lower bound should be less than or equal to the upper bound.

Constructors

Interval v v
Lower and upper bounds of the interval.

low :: Interval v -> v

Lower bound of the interval

high :: Interval v -> v

Upper bound of the interval

point :: v -> Interval v

An interval in which the lower and upper bounds are equal.

newtype I a

The identity type functor.

Like Identity, but with a shorter name.

Constructors

I a

unI :: I a -> a

Extract the contents of an I value.

module Data.SOP.Strict

chunksOf :: Int -> [e] -> [[e]]

chunksOf n splits a list into length-n pieces. The last piece will be shorter if n does not evenly divide the length of the list. If n <= 0, chunksOf n l returns an infinite list of empty lists.

>>> chunksOf 3 [1..12]
[[1,2,3],[4,5,6],[7,8,9],[10,11,12]]

>>> chunksOf 3 "Hello there"
["Hel","lo ","the","re"]

>>> chunksOf 3 ([] :: [Int])
[]

Note that chunksOf n [] is [], not [[]]. This is intentional, and satisfies the property that

chunksOf n xs ++ chunksOf n ys == chunksOf n (xs ++ ys)

whenever n evenly divides the length of xs.

data UTCTime

This is the simplest representation of UTC. It consists of the day number, and a time offset from midnight. Note that if a day has a leap second added to it, it will have 86401 seconds.

Constructors

UTCTime

Fields

utctDay :: Day

the day

utctDayTime :: DiffTime

the time from midnight, 0 <= t < 86401s (because of leap-seconds)

data NominalDiffTime

This is a length of time, as measured by UTC. It has a precision of 10^-12 s.

Conversion functions such as fromInteger and realToFrac will treat it as seconds. For example, (0.010 :: NominalDiffTime) corresponds to 10 milliseconds.

It has a precision of one picosecond (= 10^-12 s). Enumeration functions will treat it as picoseconds.

It ignores leap-seconds, so it's not necessarily a fixed amount of clock time. For instance, 23:00 UTC + 2 hours of NominalDiffTime = 01:00 UTC (+ 1 day), regardless of whether a leap-second intervened.

addUTCTime :: NominalDiffTime -> UTCTime -> UTCTime

addUTCTime a b = a + b

diffUTCTime :: UTCTime -> UTCTime -> NominalDiffTime

diffUTCTime a b = a - b

module Data.Time.Clock.POSIX

data StrictMaybe a

Constructors

SNothing

SJust !a

fromSMaybe :: a -> StrictMaybe a -> a

Same as fromMaybe

(***) :: Arrow a => a b c -> a b' c' -> a (b, b') (c, c') infixr 3

Split the input between the two argument arrows and combine their output. Note that this is in general not a functor.

The default definition may be overridden with a more efficient version if desired.

(&&&) :: Arrow a => a b c -> a b c' -> a b (c, c') infixr 3

Fanout: send the input to both argument arrows and combine their output.

The default definition may be overridden with a more efficient version if desired.

(<|>) :: Alternative f => f a -> f a -> f a infixl 3

An associative binary operation

mapConcurrently :: Traversable t => (a -> IO b) -> t a -> IO (t b)

forConcurrently :: Traversable t => t a -> (a -> IO b) -> IO (t b)

mapConcurrently_ :: Foldable f => (a -> IO b) -> f a -> IO ()

forConcurrently_ :: Foldable f => f a -> (a -> IO b) -> IO ()

newExceptT :: m (Either x a) -> ExceptT x m a

Constructor for computations in the ExceptT monad. (The inverse of runExceptT).

firstExceptT :: forall (m :: Type -> Type) x y a. Functor m => (x -> y) -> ExceptT x m a -> ExceptT y m a

Map the Left unwrapped computation using the given function.

module Ouroboros.Consensus.Util.Time

printf :: PrintfType r => String -> r

Format a variable number of arguments with the C-style formatting string.

>>> printf "%s, %d, %.4f" "hello" 123 pi
hello, 123, 3.1416

The return value is either String or (IO a) (which should be (IO ()), but Haskell's type system makes this hard).

The format string consists of ordinary characters and conversion specifications, which specify how to format one of the arguments to printf in the output string. A format specification is introduced by the % character; this character can be self-escaped into the format string using %%. A format specification ends with a format character that provides the primary information about how to format the value. The rest of the conversion specification is optional. In order, one may have flag characters, a width specifier, a precision specifier, and type-specific modifier characters.

Unlike C printf(3), the formatting of this printf is driven by the argument type; formatting is type specific. The types formatted by printf "out of the box" are:

Integral types, including Char

String

RealFloat types

printf is also extensible to support other types: see below.

A conversion specification begins with the character %, followed by zero or more of the following flags:

-      left adjust (default is right adjust)
+      always use a sign (+ or -) for signed conversions
space  leading space for positive numbers in signed conversions
0      pad with zeros rather than spaces
#      use an \"alternate form\": see below

When both flags are given, - overrides 0 and + overrides space. A negative width specifier in a * conversion is treated as positive but implies the left adjust flag.

The "alternate form" for unsigned radix conversions is as in C printf(3):

%o           prefix with a leading 0 if needed
%x           prefix with a leading 0x if nonzero
%X           prefix with a leading 0X if nonzero
%b           prefix with a leading 0b if nonzero
%[eEfFgG]    ensure that the number contains a decimal point

Any flags are followed optionally by a field width:

num    field width
*      as num, but taken from argument list

The field width is a minimum, not a maximum: it will be expanded as needed to avoid mutilating a value.

Any field width is followed optionally by a precision:

.num   precision
.      same as .0
.*     as num, but taken from argument list

Negative precision is taken as 0. The meaning of the precision depends on the conversion type.

Integral    minimum number of digits to show
RealFloat   number of digits after the decimal point
String      maximum number of characters

The precision for Integral types is accomplished by zero-padding. If both precision and zero-pad are given for an Integral field, the zero-pad is ignored.

Any precision is followed optionally for Integral types by a width modifier; the only use of this modifier being to set the implicit size of the operand for conversion of a negative operand to unsigned:

hh     Int8
h      Int16
l      Int32
ll     Int64
L      Int64

The specification ends with a format character:

c      character               Integral
d      decimal                 Integral
o      octal                   Integral
x      hexadecimal             Integral
X      hexadecimal             Integral
b      binary                  Integral
u      unsigned decimal        Integral
f      floating point          RealFloat
F      floating point          RealFloat
g      general format float    RealFloat
G      general format float    RealFloat
e      exponent format float   RealFloat
E      exponent format float   RealFloat
s      string                  String
v      default format          any type

The "%v" specifier is provided for all built-in types, and should be provided for user-defined type formatters as well. It picks a "best" representation for the given type. For the built-in types the "%v" specifier is converted as follows:

c      Char
u      other unsigned Integral
d      other signed Integral
g      RealFloat
s      String

Mismatch between the argument types and the format string, as well as any other syntactic or semantic errors in the format string, will cause an exception to be thrown at runtime.

Note that the formatting for RealFloat types is currently a bit different from that of C printf(3), conforming instead to showEFloat, showFFloat and showGFloat (and their alternate versions showFFloatAlt and showGFloatAlt). This is hard to fix: the fixed versions would format in a backward-incompatible way. In any case the Haskell behavior is generally more sensible than the C behavior. A brief summary of some key differences:

Haskell printf never uses the default "6-digit" precision used by C printf.

Haskell printf treats the "precision" specifier as indicating the number of digits after the decimal point.

Haskell printf prints the exponent of e-format numbers without a gratuitous plus sign, and with the minimum possible number of digits.

Haskell printf will place a zero after a decimal point when possible.

type SMaybe a = StrictMaybe a

data F

Constructors

R String

RNoCR String

Q String

L [String]

forall a.ToJSON a => J a

newtype RUTCTime

Constructors

RUTCTime

Fields

unRUTCTime :: NominalDiffTime

count :: (a -> Bool) -> [a] -> Int

isSNothing :: SMaybe a -> Bool

isSJust :: SMaybe a -> Bool

strictMaybe :: Maybe a -> SMaybe a

replaceExtension :: FilePath -> String -> FilePath

showText :: Show a => a -> Text

sequenceConcurrentlyChunksOf :: Int -> [IO a] -> IO [a]

withTimingInfo :: MonadIO m => String -> m a -> m a

spans :: forall a. (a -> Bool) -> [a] -> [Vector a]

unionIntv :: Ord a => [Interval a] -> Interval a

intersectIntv :: Ord a => [Interval a] -> Interval a

renderIntv :: (a -> Text) -> Interval a -> Text

intvDurationSec :: Interval UTCTime -> NominalDiffTime

smaybe :: b -> (a -> b) -> StrictMaybe a -> b

lazySMaybe :: SMaybe a -> Maybe a

catSMaybes :: [SMaybe a] -> [a]

mapSMaybe :: (a -> StrictMaybe b) -> [a] -> [b]

mapSMaybeFB :: (b -> r -> r) -> (a -> StrictMaybe b) -> a -> r -> r

mapConcurrentlyPure :: NFData b => (a -> b) -> [a] -> IO [b]

mapHead :: (a -> a) -> [a] -> [a]

mapLast :: (a -> a) -> [a] -> [a]

redistribute :: (a, (b, c)) -> ((a, b), (a, c))

nChunksEachOf :: Int -> Int -> Text -> [Text]

toDouble :: forall a. Real a => a -> Double

progressLock :: Lock

progress :: MonadIO m => String -> F -> m ()

norm2Tuple :: ((a, b), c) -> (a, (b, c))

roundUTCTimeSec :: UTCTime -> UTCTime

roundUTCTimeDay :: UTCTime -> UTCTime

utcTimeDeltaSec :: UTCTime -> UTCTime -> Int

foldEmpty :: r -> ([a] -> r) -> [a] -> r

zeroUTCTime :: UTCTime

A tweaked version of UTCTime that is able to have more instances. Structurally equivalent to difftime from zeroUTCTime

unsafeNominalToUTC :: NominalDiffTime -> UTCTime

unsafeUTCToNominal :: UTCTime -> NominalDiffTime

toRUTCTime :: UTCTime -> RUTCTime

fromRUTCTime :: RUTCTime -> UTCTime

mapAndUnzip :: (a -> (b, c)) -> [a] -> ([b], [c])

Orphan instances

Functor Interval

Instance details

Methods

fmap :: (a -> b) -> Interval a -> Interval b

(<$) :: a -> Interval b -> Interval a

FromJSON a => FromJSON (Interval a)

Instance details

Methods

parseJSON :: Value -> Parser (Interval a)

parseJSONList :: Value -> Parser [Interval a]

omittedField :: Maybe (Interval a)

FromJSON a => FromJSON (I a)

Instance details

Methods

parseJSON :: Value -> Parser (I a)

parseJSONList :: Value -> Parser [I a]

omittedField :: Maybe (I a)

ToJSON a => ToJSON (Interval a)

Instance details

Methods

toJSON :: Interval a -> Value

toEncoding :: Interval a -> Encoding

toJSONList :: [Interval a] -> Value

toEncodingList :: [Interval a] -> Encoding

omitField :: Interval a -> Bool

ToJSON a => ToJSON (I a)

Instance details

Methods

toJSON :: I a -> Value

toEncoding :: I a -> Encoding

toJSONList :: [I a] -> Value

toEncodingList :: [I a] -> Encoding

omitField :: I a -> Bool

Data a => Data (SMaybe a)

Instance details

Methods

gfoldl :: (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> SMaybe a -> c (SMaybe a)

gunfold :: (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> Constr -> c (SMaybe a)

toConstr :: SMaybe a -> Constr

dataTypeOf :: SMaybe a -> DataType

dataCast1 :: Typeable t => (forall d. Data d => c (t d)) -> Maybe (c (SMaybe a))

dataCast2 :: Typeable t => (forall d e. (Data d, Data e) => c (t d e)) -> Maybe (c (SMaybe a))

gmapT :: (forall b. Data b => b -> b) -> SMaybe a -> SMaybe a

gmapQl :: (r -> r' -> r) -> r -> (forall d. Data d => d -> r') -> SMaybe a -> r

gmapQr :: forall r r'. (r' -> r -> r) -> r -> (forall d. Data d => d -> r') -> SMaybe a -> r

gmapQ :: (forall d. Data d => d -> u) -> SMaybe a -> [u]

gmapQi :: Int -> (forall d. Data d => d -> u) -> SMaybe a -> u

gmapM :: Monad m => (forall d. Data d => d -> m d) -> SMaybe a -> m (SMaybe a)

gmapMp :: MonadPlus m => (forall d. Data d => d -> m d) -> SMaybe a -> m (SMaybe a)

gmapMo :: MonadPlus m => (forall d. Data d => d -> m d) -> SMaybe a -> m (SMaybe a)

NFData a => NFData (Interval a)

Instance details

Methods

rnf :: Interval a -> ()